requirements.txt
pandas==1.5.1
python-dotenv==0.21.1
